pyproject.toml
[tool.poetry]
name = "podcast-llm"
version = "0.2.1"
description = "An intelligent system that automatically generates engaging podcast conversations using LLMs and text-to-speech technology."
authors = ["Evan Dempsey <evandempsey@gmail.com>"]
license = "Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)"
readme = "README.md"
repository = "https://github.com/evandempsey/podcast-llm"
documentation = "https://evandempsey.github.io/podcast-llm/"
keywords = ["podcast", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License", 
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.4"
langchain-anthropic = "^0.3.0"
langchain-openai = "^0.2.8"
pydub = "^0.25.1"
google-cloud-texttospeech = "^2.21.1"
elevenlabs = "^1.12.1"
langchain-community = "^0.3.7"
tavily-python = "^0.5.0"
youtube-transcript-api = "^0.6.2"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.4.1"
pypdf = "^5.1.0"
openai = "^1.54.4"
wikipedia = "^1.4.0"
audioop-lts = { version = "^0.2.0", python = "^3.13" }
gradio = "^5.6.0"
gradio-log = "^0.0.7"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
sphinxcontrib-napoleon = "^0.7"

[tool.poetry.scripts]
podcast-llm = "podcast_llm.generate:main"
podcast-llm-gui = "podcast_llm.gui:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# This ensures your package is importable
packages = [
    { include = "podcast_llm" }
]
